Completed Auth, Profile, Address, Shopper modules with Cloudinary upload

+++ src/main/java/com/ecommerce/authdemo/repository/AddressRepository.java
@@ -12,7 +12,5 @@ public interface AddressRepository extends JpaRepository<Address, Integer> {
 
     Optional<Address> findByUserIdAndIsDefaultTrue(Integer userId);
 
-    Optional<Address> findByIdAndUserId(Long id, Long userId);
-
-
+    Optional<Address> findByIdAndUserId(Integer id, Integer userId);
 }
